edumentor/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
groq>=0.13.0
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.1
python-pptx==0.6.21
gunicorn==21.2.0
